requirements.txt
numpy
pandas
seaborn
torch
tqdm
matplotlib
yellowbrick
omegaconf
pathlib
hydra-core